+++ libs/utils/src/lib.rs
@@ -2,8 +2,6 @@
 //! between other crates in this repository.
 #![deny(clippy::undocumented_unsafe_blocks)]
 
-extern crate hyper0 as hyper;
-
 pub mod backoff;
 
 /// `Lsn` type implements common tasks on Log Sequence Numbers
@@ -32,9 +30,6 @@ pub mod shard;
 
 mod hex;
 pub use hex::Hex;
-
-// http endpoint utils
-pub mod http;
 
 // definition of the Generation type for pageserver attachment APIs
 pub mod generation;
@@ -95,8 +90,6 @@ pub mod toml_edit_ext;
 pub mod circuit_breaker;
 
 pub mod try_rcu;
-
-pub mod pprof;
 
 pub mod guard_arc_swap;
